wgml: do code more portable
remove/change 16-bit support residual code and data

+++ bld/wgml1/c/gdata.c
@@ -43,7 +43,7 @@ void init_global_vars( void )
     master_fname        = NULL;         // Master input file name
     master_fname_attr   = NULL;         // Master input file name attributes
     line_from           = 1;            // default first line to process
-    line_to             = ULONG_MAX -1; // default last line to process
+    line_to             = UINT_MAX - 1; // default last line to process
     dev_name            = NULL;         // device defined_name
     opt_fonts           = NULL;         // option FONT parameters (linked list)
